[SYNTH-12989] Unexport `runTests()`
